requirements.txt
django>=4.2
djangorestframework>=3.14
django-cors-headers
django-filter
djangorestframework-simplejwt
python-decouple
